requirements.txt
# greenlet 需指定有 Windows 预编译轮的版本，避免本地编译
greenlet==3.0.3
Flask>=2.3.0
Flask-Login>=0.6.2
Flask-SQLAlchemy>=3.0.5
PyMySQL>=1.1.0
bcrypt>=4.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
